extract data fetching into a custom hook
