+++ src/features/products/components/ProductGallery/ProductGallery.tsx
@@ -3,31 +3,16 @@ import { ProductGalleryItem } from "../ProductGalleryItem/ProductGalleryItem.tsx
 import { useInfiniteQuery } from "@tanstack/react-query";
 import { getProducts, Product } from "../../../../services/products.service.ts";
 import { VirtuosoGrid } from "react-virtuoso";
+import { useProductsQueryData } from "../../hooks/useProductsQueryData.tsx";
 
 const LIMIT = 10;
 const SKIP = 10;
 
 export const ProductGallery = () => {
-  const { data, isFetching, fetchNextPage } = useInfiniteQuery({
-    queryKey: ["products"],
-    queryFn: getProducts,
-    initialPageParam: { limit: LIMIT, skip: 0 },
-    getNextPageParam: (lastPage) => ({
-      limit: LIMIT,
-      skip: lastPage.skip + SKIP,
-    }),
-  });
-
-  const productData = useMemo(() => {
-    return (
-      data?.pages.reduce<Product[]>(
-        (products, page) => products.concat(page.products),
-        [],
-      ) ?? []
-    );
-  }, [data]);
-
-  const totalCount = useMemo(() => data?.pages[0]?.total, [data]);
+  const { data, isFetching, totalCount, fetchNextPage } = useProductsQueryData(
+    LIMIT,
+    SKIP,
+  );
 
   return (
     <VirtuosoGrid
@@ -37,7 +22,7 @@ export const ProductGallery = () => {
         maxWidth: "1280px",
       }}
       listClassName="grid gap-4 md:grid-cols-3 sm:grid-cols-2 justify-center max-w-screen-xl"
-      data={productData}
+      data={data}
       totalCount={totalCount}
       endReached={() => fetchNextPage()}
       increaseViewportBy={200}
